Add exception tratament

+++ app/services/category_services.ts
@@ -1,3 +1,5 @@
+import ConflictException from '#exceptions/conflict_exception'
+import NotFoundException from '#exceptions/not_found_exception'
 import Category from '#models/category'
 import { DateTime } from 'luxon'
 
@@ -12,12 +14,21 @@ interface UpdateCategoryData {
 
 export default class CategoryServices {
   async createCategory({ name }: CreateCategoryData) {
+    const verifyCategory = await Category.findBy('name', name)
+    if (verifyCategory) {
+      throw new ConflictException('Categoria já cadastrada')
+    }
+
     const category = await Category.create({ name })
     return category
   }
 
   async updateCategory({ id, name }: UpdateCategoryData) {
-    const category = await Category.findByOrFail('id', id)
+    const category = await Category.findBy('id', id)
+    if (!category) {
+      throw new NotFoundException('Categoria não encontrada')
+    }
+
     category.name = name
     category.updatedAt = DateTime.now()
 
@@ -26,17 +37,26 @@ export default class CategoryServices {
   }
 
   async getById(id: number) {
-    const category = await Category.findByOrFail('id', id)
+    const category = await Category.findBy('id', id)
+    if (!category) {
+      throw new NotFoundException('Categoria não encontrada')
+    }
     return category
   }
 
   async getByName(name: string) {
-    const category = await Category.findByOrFail('name', name)
+    const category = await Category.findBy('name', name)
+    if (!category) {
+      throw new NotFoundException('Categoria não encontrada')
+    }
     return category
   }
 
   async deleteCategory(id: number) {
-    const category = await Category.findByOrFail('id', id)
+    const category = await Category.findBy('id', id)
+    if (!category) {
+      throw new NotFoundException('Categoria não encontrada')
+    }
 
     await category.delete()
   }
